requirements.txt
strands-agents
strands-agents-tools
bedrock-agentcore
boto3
streamlit
streamlit-authenticator